pyproject.toml
[project]
name = "unet"
version = "0.1.0"
description = "An exploration of the paper _A Unified Framework for U-Net Design and Analysis, by Williams, Falck, et al._"
readme = "README.md"

classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.12",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License"
]

authors = [{ name = "Pierre Fontbonne", email = "pierre.fontbonne@etu.sorbonne-universite.fr" }]
maintainers = [
    { name = "Pierre Fontbonne", email = "pierre.fontbonne@etu.sorbonne-universite.fr" },
]
license = "MIT"

requires-python = ">= 3.12"

dependencies = [
    "hydra-core>=1.3.2",
    "pytorch-lightning>=2.6.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[project.urls]
Homepage = "https://github.com/fntb/unet"
Documentation = "https://github.com/fntb/unet"
Repository = "https://github.com/fntb/unet"
Issues = "https://github.com/fntb/unet"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "src",
    "src.conf",
    "src.conf.data",
    "src.conf.experiment",
    "src.conf.model",
    "src.conf.trainer",
    "src.data",
    "src.metrics",
    "src.models",
    "src.optim",
]
